Move stem class to separate file
