+++ src/beam.cpp
@@ -19,7 +19,6 @@
 #include "btrem.h"
 #include "doc.h"
 #include "editorial.h"
-#include "elementpart.h"
 #include "functorparams.h"
 #include "gracegrp.h"
 #include "layer.h"
@@ -29,6 +28,7 @@
 #include "smufl.h"
 #include "space.h"
 #include "staff.h"
+#include "stem.h"
 #include "tabdursym.h"
 #include "tabgrp.h"
 #include "tuplet.h"
