+++ src/drawinginterface.cpp
@@ -14,11 +14,11 @@
 //----------------------------------------------------------------------------
 
 #include "chord.h"
-#include "elementpart.h"
 #include "layerelement.h"
 #include "note.h"
 #include "object.h"
 #include "staff.h"
+#include "stem.h"
 
 namespace vrv {
