+++ src/view_element.cpp
@@ -54,6 +54,7 @@
 #include "smufl.h"
 #include "space.h"
 #include "staff.h"
+#include "stem.h"
 #include "syl.h"
 #include "system.h"
 #include "textelement.h"
